🚨Fix: Notice Item prop 변경, 알림 원본 페이지 이동 경로 수정
typeToMessage 함수 분리
댓글 미리보기 스타일 수정
